Implemented getQuarter() api call
